+++ lib/src/environment.dart
@@ -3,6 +3,9 @@ import 'dart:convert' show jsonDecode;
 import 'package:flutter/services.dart' show rootBundle;
 import 'package:flutter_app_environment/src/models/debug_options.dart';
 import 'package:flutter_app_environment/src/models/environment_type.dart';
+import 'package:flutter_app_environment/src/models/exceptions/environment_already_initialized_exception.dart';
+import 'package:flutter_app_environment/src/models/exceptions/environment_failed_to_load_exception.dart';
+import 'package:flutter_app_environment/src/models/exceptions/environment_not_initialized_exception.dart';
 
 class Environment<T> {
   const Environment._({
@@ -15,7 +18,7 @@ class Environment<T> {
 
   factory Environment.instance() {
     if (_instance == null) {
-      throw Exception('Need call init method first!');
+      throw const EnvironmentNotInitializedException();
     }
 
     return _instance! as Environment<T>;
@@ -32,7 +35,7 @@ class Environment<T> {
   /// Throws an [Exception] when Environment.init not called before and Environment not initialized.
   Enum get currentEnvironmentType {
     if (_instance == null) {
-      throw Exception('Need call init method first!');
+      throw const EnvironmentNotInitializedException();
     }
 
     return _currentEnvironmentType;
@@ -43,7 +46,7 @@ class Environment<T> {
   /// Throws an [Exception] when Environment.init not called before and Environment not initialized.
   T get config {
     if (_instance == null) {
-      throw Exception('Need call init method first!');
+      throw const EnvironmentNotInitializedException();
     }
 
     return _config;
@@ -54,7 +57,7 @@ class Environment<T> {
   /// Throws an [Exception] when Environment.init not called before and Environment not initialized.
   DebugOptions get debugOptions {
     if (_instance == null) {
-      throw Exception('Need call init method first!');
+      throw const EnvironmentNotInitializedException();
     }
 
     return _debugOptions;
@@ -65,7 +68,7 @@ class Environment<T> {
   /// Throws an [Exception] when Environment.init not called before and Environment not initialized.
   bool get isDebug {
     if (_instance == null) {
-      throw Exception('Need call init method first!');
+      throw const EnvironmentNotInitializedException();
     }
 
     return _currentEnvironmentType == EnvironmentType.development;
@@ -91,7 +94,9 @@ class Environment<T> {
         config: config,
       );
     } else {
-      throw Exception('Environment already initialized');
+      throw EnvironmentAlreadyInitializedException(
+        'Environment already initialized with ${_instance!.currentEnvironmentType} environment type',
+      );
     }
   }
 
@@ -115,7 +120,9 @@ class Environment<T> {
         config: config,
       );
     } else {
-      throw Exception('Environment already initialized');
+      throw EnvironmentAlreadyInitializedException(
+        'Environment already initialized with ${_instance!.currentEnvironmentType} environment type',
+      );
     }
   }
 
@@ -123,7 +130,7 @@ class Environment<T> {
   ///
   /// Params:
   ///   [environmentType] The environment type that you want to initialize the environment with.
-  ///   [fromJson] A function that takes a Map<String, dynamic> from environment file and returns an instance of the config class.
+  ///   [fromJson] A function that takes a Map\<String, dynamic\> from environment file and returns an instance of the config class.
   ///   [debugOptions] An optional parameter that allows you to specify the debug options for the environment.
   ///
   /// Throws an [Exception] when [Environment] already initialized.
@@ -136,16 +143,25 @@ class Environment<T> {
       final environmentFileName = '${environmentType.toShortString()}.json';
       final environmentPath = 'res/config/$environmentFileName';
 
-      final content = await rootBundle.loadString(environmentPath);
-      final jsonObject = jsonDecode(content) as Map<String, dynamic>;
+      try {
+        final content = await rootBundle.loadString(environmentPath);
+        final jsonObject = jsonDecode(content) as Map<String, dynamic>;
 
-      _instance = Environment<T>._(
-        environmentType: environmentType,
-        debugOptions: debugOptions,
-        config: fromJson(jsonObject),
-      );
+        _instance = Environment<T>._(
+          environmentType: environmentType,
+          debugOptions: debugOptions,
+          config: fromJson(jsonObject),
+        );
+        // ignore: avoid_catching_errors
+      } on Error {
+        throw EnvironmentFailedToLoadException(
+          'Failed to load environment file: $environmentPath',
+        );
+      }
     } else {
-      throw Exception('Environment already initialized');
+      throw EnvironmentAlreadyInitializedException(
+        'Environment already initialized with ${_instance!.currentEnvironmentType} environment type',
+      );
     }
   }
 
@@ -153,7 +169,7 @@ class Environment<T> {
   ///
   /// Params:
   ///   [environmentType] The environment type that you want to initialize the environment with.
-  ///   [fromJson] A function that takes a Map<String, dynamic> from environment file and returns an instance of the config class.
+  ///   [fromJson] A function that takes a Map\<String, dynamic\> from environment file and returns an instance of the config class.
   ///   [debugOptions] An optional parameter that allows you to specify the debug options for the environment.
   ///
   /// Throws an [Exception] when [Environment] already initialized.
@@ -175,7 +191,9 @@ class Environment<T> {
         config: fromJson(jsonObject),
       );
     } else {
-      throw Exception('Environment already initialized');
+      throw EnvironmentAlreadyInitializedException(
+        'Environment already initialized with ${_instance!.currentEnvironmentType} environment type',
+      );
     }
   }
 }