+++ lib/src/models/exceptions/environment_already_initialized_exception.dart
@@ -1,0 +1,12 @@
+/// Exception that is thrown when the environment is already initialized.
+class EnvironmentAlreadyInitializedException implements Exception {
+  const EnvironmentAlreadyInitializedException(this.message);
+
+  /// The message of the exception.
+  final String message;
+
+  @override
+  String toString() {
+    return 'EnvironmentAlreadyInitializedException: $message';
+  }
+}